requirements.txt
fastapi
uvicorn
mangum
langchain
langgraph
langchain-openai
python-dotenv
boto3
kubernetes
requests
